cpu: lpc2387: integrated lpc_common into lpc2387
Rationale 1: The common part made only sense for (some) NXP ARM7 MCUs,
but was misleading for MCUs like the LPC1768.

Rationale 2: The common part was only used by one specific MCU
implementation - no need to outsource it.
