+++ cpu/lpc2387/hwtimer_cpu.c
