+++ cpu/lpc2387/lpc_syscalls.c
