--- cpu/lpc_common/Makefile
@@ -1,1 +1,0 @@
-include $(RIOTBASE)/Makefile.base
