--- cpu/lpc_common/Makefile.include
@@ -1,3 +1,0 @@
-INCLUDES += -I$(RIOTCPU)/lpc_common/include
-
-export UNDEF += $(BINDIR)lpc_common/lpc_syscalls.o
