--- cpu/lpc_common/doc.txt
@@ -1,13 +1,0 @@
-/*
- * Copyright (C) 2013 Freie Universität Berlin
- *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
- */
-
-/**
- * @defgroup    lpc_common LPC common
- * @brief       Common code for all arm-based LPC controllers
- * @ingroup     cpu
- */
