+++ cpu/msp430fxyz/Makefile
@@ -1,0 +1,7 @@
+MODULE =cpu
+
+include $(RIOTCPU)/$(CPU)/Makefile.include
+
+DIRS = $(RIOTCPU)/msp430-common/
+
+include $(RIOTBASE)/Makefile.base
